DRY blob db tests
